Table and packet updates
Table now takes pdu_header struct. This includes
the sequence number (in host order) and the actual pdu.
Ideally the table will work as a window but it needs
more testing. It works as a circular queue with a "pseudo window."
The "pseudo window" is fancy for saying that the current, upper, and
lower values increase outside the scope of the actual data structure.

Packets with the data flag can be built. The checksum is created
and successfully passes and fails on good and bad packets
respectively.